NoIntents handling
-NoIntents handling - when you say no during date or time confirmation
starts a new booking
-"Sorry, there is no flight from {} to {} on None" bug fixed. Now Alexa
asks for date
