--- BookingBot.Services.Skill/README.md
@@ -1,1 +1,0 @@
-# BookingBot
