--- BookingBot.Services.Skill/amazon_portal_values/Utterances.txt
@@ -1,52 +1,0 @@
-DestinationPlaceIntent to {name}
-DestinationPlaceIntent I want to go to {name}
-DestinationPlaceIntent I want to go to the {name}
-DestinationPlaceIntent I want to fly to {name}
-DestinationPlaceIntent I would like to go to {name}
-DestinationPlaceIntent I would like to go to the {name}
-DestinationPlaceIntent I would like to fly to {name}
-
-DeparturePlaceIntent from {name}    
-DeparturePlaceIntent I will fly from {name}
-
-PlaceIntent {name}
-
-DepartureDateIntent {the_date}
-DepartureDateIntent I want to fly {the_date}
-DepartureDateIntent I want to fly at {the_date}
-DepartureDateIntent I would like to fly on {the_date}
-DepartureDateIntent I would like to fly {the_date}
-DepartureDateIntent Find me flight at {the_date}
-DepartureDateIntent Find me flight {the_date}
-
-ChooseFlightIntent Book me flight at {departure_time}
-ChooseFlightIntent Yes book me flight at {departure_time}
-ChooseFlightIntent Yes book me flight on {departure_time}
-ChooseFlightIntent Yes flight at {departure_time}
-ChooseFlightIntent Yes flight on {departure_time}
-ChooseFlightIntent Yes {departure_time}
-ChooseFlightIntent Yes at {departure_time}
-ChooseFlightIntent at {departure_time}
-ChooseFlightIntent {departure_time}
-
-HowManySeatsToBook {number}
-
-BookingIntent I would like to fly to {destination_city} from {departure_city} on {departure_date}
-BookingIntent I want to fly to {destination_city} from {departure_city} on {departure_date}
-BookingIntent I would like to fly from {departure_city} to {destination_city} on {departure_date}
-BookingIntent I want to fly from {departure_city} to {destination_city} on {departure_date}
-
-BookingIntent I would like to fly on {departure_date} to {destination_city} from {departure_city} 
-BookingIntent I want to fly {departure_date} to {destination_city} from {departure_city} 
-BookingIntent I would like to fly on {departure_date} from {departure_city} to {destination_city} 
-BookingIntent I want to fly {departure_date} from {departure_city} to {destination_city} 
-
-BookingIntent Book me a flight to {destination_city} from {departure_city} on {departure_date}
-BookingIntent Book me a flight to {destination_city} from {departure_city} {departure_date}
-BookingIntent Book me a flight from {departure_city} to {destination_city} on {departure_date}
-BookingIntent Book me a flight from {departure_city} to {destination_city} {departure_date}
-
-BookingIntent Book me a flight on {departure_date} to {destination_city} from {departure_city}
-BookingIntent Book me a flight {departure_date} to {destination_city} from {departure_city} 
-BookingIntent Book me a flight on {departure_date} from {departure_city} to {destination_city} 
-BookingIntent Book me a flight {departure_date} from {departure_city} to {destination_city}
