--- BookingBot.Services.Skill/templates.yaml
@@ -1,86 +1,0 @@
-welcome: |
-    Welcome to Ryanair booking system. Where would you like to fly?
-chooseFlightAgain: |
-    Ok, lets go through that again. Where would you like to fly?
-
-#place templates
-noSuchDeparture: |
-    Sorry, we don't flight from {}
-noSuchDestination: |
-    Sorry, we don't flight to {}
-noSuchDeaprturePlace_ChooseAnother: |
-    Sorry, we don't operate in {}. Please choose another departure place'
-noSuchDestinationPlace_ChooseAnother: |
-    Sorry, we don't operate in {}. Please choose another destination'
-noFlightConnection: |
-    Sorry, we don't fly from {} to {}
-askForDeparturePlace: |
-    Where will you travel from?
-destinationAndDepartureCollected: |
-    Great, when do you want to flight from {} to {}
-
-#date templates
-foundFlight: |
-  <speak> 
-    Ok, the flight from {} to {} on {} at {}.
-     <break time="1s"/> 
-    Would you like to book it? 
-  </speak>
-
-foundFlightsBeginning: |
-  <speak> 
-    There are {} flights from {} to {} on {}. 
-    <break time="1s"/> 
-foundFlightsMiddle: |
-    Departure at {}. 
-    Arrival at {} 
-    <break time="1s"/> 
-foundFlightsEnd: |
-    Would you like to book one of these? </speak> 
-
-noSuchFlightAtDate: | 
-  <speak> 
-    Sorry, there is no flight from {} to {} on {}.
-    When would you like to fly? 
-  </speak>
-
-#time templates
-askForFlightTime: |
-    Which flight hour do you choose?
-askToRepeatTime: |
-    There is no flight at {}. Please repeat departure time.
-flightTimeChosen: |
-    Flight on {} at {}. Are you sure?
-
-#more_info templates
-askForSeatsAmount: |
-    How many seats do you want to book?
-askToRepeatAmount: |
-    Sorry, I didn't get it. How many seats do you want to book?
-askForSeat: |
-    Do you want to choose a seat for ten euros?
-askToRepeatSeat: |
-    Sorry, I didn't get it. Do you want to choose a seat for ten euros for each passenger?
-askForInsurance: |
-    Do you want to book insurance?
-askToRepeatInsurance: |
-    Sorry, I didn't get it. Do you want to book insurance for fifteen euros for each passenger?
-saySummaryAndConfirm: |
-    Please confirm the following booking details. You departure on {} 
-    from {} to {}. 
-saySummaryAndConfirmIncludes: |
-    Your reservation includes {} passenger
-saySummaryAndConfirmSeatReservation: |
-    seat reservation 
-saySummaryAndConfirmInsurance: |
-    insurance 
-saySummaryAndConfirmTotalPrice: |
-    The total price is {} euro. 
-saySummaryAndConfirmDoYouConfirm: |
-    Do you want to book it?
-bookingDone: |
-    Congratulations, you have just booked the flight. Thank you and have a nice flight.
-bookingCancelled: |
-    Your booking is cancelled. Goodbye
-askForLastConfirmation: |
-    Sorry, I didn't get it. Do you want to book insurance?
