--- BookingBot.Services.Skill/utils/constants/constants.py
@@ -1,20 +1,0 @@
-"""Flight info"""
-DEPARTURE_DATE = "DEPARTURE_DATE"
-DEPARTURE_TIME = "DEPARTURE_TIME"
-DEPARTURE_CITY = "DEPARTURE_CITY"
-DESTINATION_CITY = "DESTINATION_CITY"
-
-FLIGHTS_ID = "FLIGHTS_ID"
-
-"Additional booking info"
-PASSENGERS_NO = "PASSENGERS_NO"
-SEAT_RESERVATION = "SEAT_RESERVATION"
-INSURANCE = "INSURANCE"
-WILL_CONFIRM_BOOKING = "WILL_CONFIRM_BOOKING"
-
-"Constants"
-SEAT_PRICE = 10
-INSURANCE_PRICE = 15
-
-FLIGHT_MIN_PRICE = 40
-FLIGHT_MAX_PRICE = 400
