+++ intents/yesno_intents.py
@@ -8,7 +8,7 @@ from intents import moreinfo_intents
 from intents import date_intents
 from intents import time_intents
 
-def forward_yes_no(value):
+def forward_yes_no(sup, value):
     "Forwards bool value"
     if constants.DEPARTURE_DATE in session.attributes:
         "Answer for 'Would you like to book one of these (flights)?'"
@@ -31,9 +31,9 @@ def forward_yes_no(value):
             
                 return moreinfo_intents.response_seat_reservation(value)
         
-            return time_intents.flight_confirmation(value)
+            return time_intents.flight_confirmation(sup, value)
 
-        return date_intents.flights_choosing_confirmation(value)
+        return date_intents.flights_choosing_confirmation(sup, value)
 
 def handle_yesno_intents(ask, sup):
     "Moreinfo intents handler"
@@ -42,12 +42,12 @@ def handle_yesno_intents(ask, sup):
     @sup.guide
     def received_yes():
         "Forwards bool value"
-        return forward_yes_no(True)
+        return forward_yes_no(sup, True)
 
     @ask.intent("AMAZON.NoIntent")
     @sup.guide
     def received_no():
         "Forwards bool value"
-        return forward_yes_no(False)
+        return forward_yes_no(sup, False)
